working on server in Rust. Basic client storage implemented
