+++ rust-server-template/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "fastchat-server"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+tokio = { version = "1.19.2", features = ["macros", "sync", "rt-multi-thread"] }
+tokio-stream = "0.1.9"
+warp = "0.3"
+serde = {version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+futures = { version = "0.3", default-features = false }
+uuid = { version = "1.1.2", features = ["serde", "v4"] }
